+++ NEWS.md
@@ -1,4 +1,6 @@
 # khroma 1.8.0.9000
+## New functions
+* Add `ramp()` to produce a colour scheme that map a `numeric` vector to colours (allow rescaling to a specific midpoint).
 
 # khroma 1.8.0
 ## New functions